refactor(shopping): extract reusable shopping item row
Move shopping item row UI and behaviors into a dedicated widget file so the list page uses a reusable component without changing existing shopping functionality.
